requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.3.0
blinker==1.4
boto3==1.37.14
botocore==1.37.14
Brlapi==0.8.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.11
fasteners==0.14.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.0
future==0.18.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.3
importlib_metadata==8.6.1
importlib_resources==6.5.2
jeepney==0.7.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
keyring==23.5.0
kubernetes==32.0.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.21.0
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
posthog==3.21.0
protobuf==5.29.3
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPika==0.48.9
pyproject_hooks==1.2.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
systemd-python==234
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xdg==5
xkit==0.0.0
zipp==3.21.0
